pyproject.toml
[project]
name = "finsight-ai"
version = "0.1.0"
description = "Agentic Banking RAG with Qdrant + Cohere + RAGAS"
requires-python = ">=3.11"

dependencies = [
    "langchain==0.2.16",
    "langchain-core==0.2.38",
    "langchain-community==0.2.16",
    "langchain-groq==0.1.9",
    "langchain-cohere==0.2.4",
    "langchain-qdrant==0.1.4",
    "qdrant-client==1.11.1",
    "cohere==5.9.0",
    "tavily-python==0.5.0",
    "ragas==0.1.21",
    "datasets==2.21.0",
    "pandas==2.2.2",
    "streamlit==1.38.0",
    "pypdf==4.3.1",
    "python-dotenv==1.0.1",
]